[Cashtab] Deprecate getDeviceNotificationStyle
Summary: In the process of optimizing notifications it became apparent that `getDeviceNotificationStyle` was a fix for a legacy issue where notifications were clipping the top frame of the browser on mobile devices. This is no longer required as the top margin is now adequate without the need to set mobile specific `marginTop` values.

Test Plan:
npm test
npm start
Trigger the following notifications across web, extension and mobile (iOS Safari and android chrome):
- Send/Receive XEC
- Send/Receive/Burn eToken
- Create eToken and icon submission
- Message signed / copied
- Message/signature verified
- Message/signature not verified
- Register alias
- Error

Reviewers: bytesofman, #bitcoin_abc

Reviewed By: bytesofman, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D15323