[Cashtab] Upgrade antd
Summary:
Upgrade antd and related dependency antd/icons

For the most part, this seems to improve Cashtab look and feel. The h3 and h4 components appear off. Added light css fixes where this upgrade broke existing css.

I was not able to color match the char count on the text input fields of the SignVerifyMsg screen. imo this is not a big deal, still looks about the same.

Background: need to do this in order to upgrade to React 18, which we need to do in order to test localforage in the useWallet hook

Test Plan:
`npm test`
`npm start` and browse the screens, check at responsiveness. We don't need 100% as-before. Need to make sure everything is usable and nothing conspicuously off (switches, inputs).

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Subscribers: emack

Differential Revision: https://reviews.bitcoinabc.org/D15353