[electrum] run black linter on scripts
Summary:
Some python scripts don't have a .py extension. This is a manual run prior to renaming the files so that they will be linted in the future.
The main entry point for the application will not get an extension, so add it to includes in  .arclint
Depends on D15362

Test Plan: `arc lint -- electrum/electrum-abc`

Reviewers: #bitcoin_abc, bytesofman

Reviewed By: #bitcoin_abc, bytesofman

Differential Revision: https://reviews.bitcoinabc.org/D15357