[electrum] run flake8 on python scripts missing the .py extension
Summary:
And fix the errors.
We will add the .py extension in a follow-up commit so that these scripts remain linted.

The main entry point for the application will not get a .py extension, so add it to `include` in .arclint.

Depends on D15358

Test Plan:
```
arc lint -- electrum/electrum-abc
```

Make sure `./electrum-abc` still runs

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien, bytesofman

Differential Revision: https://reviews.bitcoinabc.org/D15359