[monorepo] Add CONTRIBUTING.md to .prettierignore
Summary:
Users with `prettier` installed globally may autoformat some md files with unrelated changes if they save any other change (removing underlines, special formattings).

Add to `.prettierignore` as these are discovered. I don't think it's worth expanding prettier rules to accept everything we have, the point of prettier is that it is "opinionated" and does not require customization.

Test Plan: Confirm this is just adding to prettierignore. See D15367 and history for how it works.

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D15368