[linter doc] Specify prettier version to install globally
Summary:
Should be updates soon. For now though, specify version so we at least know what should be used.

Note: discovered this when I upgraded to node 20 and reinstalled prettier with `npm i -g prettier`, which gives the latest at 3.2.0, which causes `arc lint --everything` to throw changes in JS repos

Test Plan: Confirm this adds a version for prettier in CONTRIBUTING.md

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D15388