[Chronik] Move `chronik-plugin` to `chronik-plugin-impl`, add `chronik-plugin` multiplex crate
Summary:
Currently, using `chronik-plugin` in the indexer is a bit tedious, as everything has to be `#[cfg(feature = "plugins")]` gated, which is hard to maintain.

Instead, we move the current `chronik-plugin` to `chronik-plugin-impl`, which is then re-exported in a new `chronik-plugin` crate if the `plugins` feature is enabled, and otherwise it exports fallback structs.

Depends on D15394.

Test Plan: `ninja check-functional`

Reviewers: Fabien, #bitcoin_abc

Reviewed By: Fabien, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D15395