[Cashtab] Increase txHistoryCount to 20
Summary:
No issues observed from landing D15281 which bumped the wallet refresh interval from 10s to 60s. This paves the way for the number of tx history items to be increased from 10 to 20 due to the reduced API load.

The primary purpose is to serve as a simple alternative to building the complex pagination feature.

The secondary purpose of this diff is to observe any performance implications of 20 in prod, as a precursor to the eCash Social platform I'm prototyping right now.

Test Plan:
npm start
observe 20 tx history items on wallet load with no performance issues

Reviewers: #bitcoin_abc, bytesofman

Reviewed By: #bitcoin_abc, bytesofman

Subscribers: Cyclonecinta88

Differential Revision: https://reviews.bitcoinabc.org/D15403