[Cashtab] Do not auto-open the QR code scanner in the extension
Summary:
We do not want to auto-open the QR code scanner in the extension, or allow that setting.

Refactor to implement.

Test Plan:
`npm run extension`, the setting in the extension is not visible, the QR code scanner does not auto-update from Send or SendToken screens

It still will for small width devices (can test on desktop).

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Subscribers: emack

Differential Revision: https://reviews.bitcoinabc.org/D15410