[Cashtab] Helper method for initializing integration tests
Summary:
Cashtab has added automated integration testing over the past few weeks. This implementation has involved some iterative discoveries and best practices.

Now that it is up and working, it's time to optimize and standardize the tests, so that it is easier to add new cases (and we do not end up accidentally testing different implementations of tests instead of test cases).

Add a helper function that initializes a mock-chronik-client instance with the data required to sync a cashtab wallet to the eCash chain.

Implement in AppIntegrations.test.js (for now, will be extended throughout the app in later diffs)

Test Plan: `npm test`

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D15429