[depends] Don't remove the include directory for native_protobuf
Summary:
The imported target created by cmake sets the INTERFACE_INCLUDE_DIRECTORIES property of the target to the installed include directory. CMake will complain that the target is using a non existent directory if we remove it in the post processing step.

This will fix the issue from D14855, which can be reverted in a follow-up.

Test Plan:
Run the gitian builds.
Undo the changes from D14855 and run the guix build.

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D15440