[Cashtab] Do not show new wallet msg for not-new wallets
Summary:
Users reporting toggling between 'congratulations new wallet' and normal wallet screen, on wallets with tx history.

Doing some digging, discovered it happened when `parsedTxHistory` was undefined.

This occurs when a new token is added to cache. The calling function is expecting info at the key `parsedTxHistory`, but this is not being returned

Test Plan:
`npm start`, receive a tx from a tokenId not in cache, do not see the 'new wallet' screen

Note: this patch is deployed to https://cashtab-local-dev.netlify.app/ and have gone through testing with Koush

Reviewers: #bitcoin_abc, Kronkmeister

Reviewed By: #bitcoin_abc, Kronkmeister

Differential Revision: https://reviews.bitcoinabc.org/D15442