[e.cash] Update strapi scorecard data
Summary:
The strapi instance used on the get eCash page was migrated to v4.
This comes with some data structure changes to the endpoint that need to be adjusted for
Mainly all of the values are nested in an attributes object now, image structure, and there is a top level data object
Also updated the scores tests and made them easier to understand by removing the verbose mocks and instead opting
for a easy to understand mock within the test file

Test Plan:
preview the site and check /get-ecash page. Should look the same
npm run test to run the tests

Reviewers: #bitcoin_abc, bytesofman

Reviewed By: #bitcoin_abc, bytesofman

Subscribers: bytesofman

Differential Revision: https://reviews.bitcoinabc.org/D15450