[Cashtab] Remove unused state param in Config
Summary:
This is used on the OnBoarding screen to prevent validation from happening while a user enters a mnemonic, since it would just fail until it is completely entered. The implementation in OnBoarding is itself a bit hard to follow and could be improved.

But it's not doing anything here and should just be removed. Probably copypasta from its use on OnBoarding.

Test Plan:
`npm test`
`git grep dirty` and it's only in OnBoarding.js

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Subscribers: emack

Differential Revision: https://reviews.bitcoinabc.org/D15475