[Cashtab] Better helper function to clear localforage between tests
Summary:
localforage.clear() is not always clearing storage in between tests. localforage is not actively maintained.

We should probably move away from localforage. However, we need to standardize how we use storage first. For now, just use an effective clearing method.

No tests added for this function as we do not want to test the library. In practice, we will notice if it is not working as existing tests will fail.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D15516