[Cashtab] Notification instead of modal for upgrades
Summary: The modal is annoying and prevents interaction with the app until you refresh. Would be nice to go "no antd" but notification for modal is even.

Test Plan:
bump version, build and deploy in netlify, do this again so netlify has this way to update, see the notification

or, trust it:
{F7364134}

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D15521