[GUIX] Build Chronik for Linux
Summary:
This builds chronik for all the supported Linux platforms. The result is a binary that doesn't include any symbol from glibc > 2.28.
This is a step toward using GUIX for our releases and have Chronik be built by default.

Test Plan:
  contrib/guix/guix-build
Run the Chronik build and the Gitian Linux build to make sure the change in the CMake file is not breaking the build.

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Subscribers: tobias_ruck

Differential Revision: https://reviews.bitcoinabc.org/D15546