misc cleanups for symbol-check
Summary:
Remove no longer needed Xenial related symbols.

The other change is a follow-up to D13971 and D13892: `check_PE_subsystem_version` does not take `filename` as an argument, and the current code only works by accident, because `filename` happens to be defined to the correct value in the outer scope when this function is called.

This is partially a backport of core#25392

Test Plan:
gitian builds
`contrib/guix/guix-build`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D15578