[Cashtab] Replace conflicting functions with more comprehensive validation function
Summary:
We should only have cashtab wallet validation function. We have two because there were in the past edge cases where a wallet was "valid", as in, it would work in Cashtab, but we still wanted to migrate it, e.g. to change the format of the address.

Implement the more comprehensive validation function everywhere.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Differential Revision: https://reviews.bitcoinabc.org/D15610