[chronik-client] Improve type definition
Summary:
Improve type definition for keys that may or may not be defined

Not expected to change behavior. However, helpful for other repos that use typescript and are using chronik-client. e.g. right now, such a repo will expect `TxInput_InNode` to have a `block` key, that may have the value of `undefined`. Reality is the `block` key may or may not exist.

discovered while working on D15625

Test Plan: CI

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Differential Revision: https://reviews.bitcoinabc.org/D15626