[cmake] Ignore empty directories during source packaging
Summary:
The patterns from the .gitignore file are excluded from the source package, but directory exclusions (like output/) will leave an empty directory in the source package.
This diff fixes this by removing the directory as well.
Also add the distsrc directories to the ignored files so guix source package is clean.

Test Plan:
Run the guix build twice and check the source package hash is deterministic.
Open the source package archive and check it is as expected.

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Subscribers: PiRK

Differential Revision: https://reviews.bitcoinabc.org/D15641