[Cashtab] Remove passLoadingStatus from components that do not require UI locking
Summary:
This prop was initially used to lock the UI when the wallet utxo set was refreshing, which could take some time (~5-10s) in the bch-api days.

Now this is much faster, and the UI is locked by useWallet.js while the wallet is being updated.

Let useWallet.js handle the UI locking for utxo set updates. Keep passLoadingStatus in the Alias and Airdrop components, where it is used to lock the UI while waiting for other server queries.

Test Plan: `npm test`, `npm start` and send some ecash and etoken txs, confirm you are not able to run into the 'invalid utxo missing or spent' error

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D15659