random: add RandomMixin::randbits with compile-known bits
Summary:
In many cases, it is known at compile time how many bits are requested from
randbits. Provide a variant of randbits that accepts this number as a template,
to make sure the compiler can make use of this knowledge. This is used immediately
in rand32() and randbool(), and a few further call sites.

This is a partial backport of [[https://github.com/bitcoin/bitcoin/pull/29625 | core#29625]]
https://github.com/bitcoin/bitcoin/pull/29625/commits/ddb7d26cfd96c1f626def4755e0e1b5aaac94d3e
Depends on D19071

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D19072