random: make GetRand() support entire range (incl. max)
Summary:
The existing code uses GetRand(nMax), with a default value for nMax, where nMax is the
range of values (not the maximum!) that the output is allowed to take. This will always
miss the last possible value (e.g. GetRand<uint32_t>() will never return 0xffffffff).

Fix this, by moving the functionality largely in RandomMixin, and also adding a
separate RandomMixin::rand function, which returns a value in the entire (non-negative)
range of an integer.

This is a partial backport of [[https://github.com/bitcoin/bitcoin/pull/29625 | core#29625]] and [[https://github.com/bitcoin/bitcoin/pull/30397 | core#30397]]
https://github.com/bitcoin/bitcoin/pull/29625/commits/e2d1f84858485650ff743753ffa5c679f210a992
Depends on D19091

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D19092