[Cashtab] Add and implement new helper function to get user locale
Summary:
`navigator.language` is a useful variable for rendering numbers and other output according to a user's browser settings and location. Add a helper function to get this variable from screens. Implement in SendXec.

This is a primitive that will be used to standardize how large numbers are formatted for Cashtab users. See D15662 for how it will first be extended and applied.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D15671